Skill 관련 Enum 파일 분리
