--- Source/PGActor/Handler/PGHandler.cpp
@@ -1,2 +1,0 @@
-﻿#include "C:\UsingProject\UnrealProject\UPlayground\Intermediate\Build\Win64\x64\UPlaygroundEditor\Development\UnrealEd\SharedPCH.UnrealEd.Project.ValApi.Cpp20.h"
-#include "PGHandler.h"
